Final implementation of JWT authentication.
Implemented:
 - Almost complete typing of current requests
 - Involves bcrypt library, for encrypting passwords
 - Added data validation via class-validator

Issue:
The code works fine, but the implementation of authentication via JWT (Passport), created a problem between class-validator and UseGuards(). When the login endpoint is called, the data is not validated by first going to UseGuards, which deals with data that is not properly processed.

The next commit will contain the standard JWT implementation